Change the representation of a TSIG error from a byte to an int.
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@1156 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/Message.java
@@ -27,7 +27,7 @@ private List [] sections;
 private int size;
 private TSIG tsigkey;
 private TSIGRecord querytsig;
-private byte tsigerror;
+private int tsigerror;
 
 int tsigstart;
 int tsigState;
@@ -510,7 +510,7 @@ toWire(int maxLength) {
  * @param querytsig If this is a response, the TSIG from the request.
  */
 public void
-setTSIG(TSIG key, byte error, TSIGRecord querytsig) {
+setTSIG(TSIG key, int error, TSIGRecord querytsig) {
 	this.tsigkey = key;
 	this.tsigerror = error;
 	this.querytsig = querytsig;
